Move plan model into separate file
